refactor(pol): renaming components
