+++ apps/hub/src/app/validators/components/RewardAllocationConfiguration.tsx
