+++ apps/hub/src/app/validators/page.tsx
@@ -1,6 +1,9 @@
 import React from "react";
 import { type Metadata } from "next";
-import Validators from "./validators";
+import { ValidatorsBanner } from "./components/validator-banner";
+import { ValidatorsGlobalInfo } from "./components/validator-global-info";
+import ValidatorsTable from "./components/ValidatorsTable";
+import { ValidatorsPortalStatus } from "./components/ValidatorsPortalStatus";
 
 export const metadata: Metadata = {
   title: "Validators",
@@ -8,5 +11,12 @@ export const metadata: Metadata = {
 };
 
 export default async function Page() {
-  return <Validators />;
+  return (
+    <div className="flex flex-col gap-16">
+      <ValidatorsBanner />
+      <ValidatorsGlobalInfo />
+      <ValidatorsTable />
+      <ValidatorsPortalStatus />
+    </div>
+  );
 }
