+++ apps/hub/src/app/validators/validator/components/BoostContent.tsx
@@ -16,7 +16,7 @@ import BigNumber from "bignumber.js";
 import { Address, parseUnits } from "viem";
 
 import ValidatorInput from "~/components/validator-input";
-import { DelegateEnum, ImageMapEnum } from "./types";
+import { DelegateEnum, ImageMapEnum } from "../types";
 
 export const DelegateContent = ({
   validator,
