+++ apps/hub/src/app/vaults/[gaugeAddress]/components/gauge-details.tsx
@@ -46,7 +46,6 @@ const _GaugeDetails = ({ gaugeAddress }: { gaugeAddress: Address }) => {
     isValidating: rewardVaultError,
   } = useRewardVault(gaugeAddress);
 
-  const router = useRouter();
   const { account } = useBeraJs();
 
   const {
